requirements.txt
numpy>=1.20.0
Pillow>=8.0.0
matplotlib>=3.3.0
ipywidgets>=7.6.0
jupyter>=1.0.0
